requirements.txt
Flask
Pillow
openai
